added test utils to makefile

+++ utils/utils.c
@@ -125,7 +125,7 @@ static void createPointerChaseInEvictionSet(Eviction_Set *evset){
 	}
 	
 	// marker array with all indexes
-	int counter=0, valid_pick=0;
+	u64 counter=0, valid_pick=0;
 	int *marker=malloc(sizeof(int)*evset->size);
 	memset(marker, 0, evset->size*sizeof(int));
 	u64 index_next=(*rng)%evset->size, index_current=0;
